third_party/WebKit/LayoutTests/imported/web-platform-tests/html/browsers/offline/application-cache-api/api_status_idle-expected.txt

This is a testharness.js-based test.
FAIL idle status test assert_equals: cache.status should equals cache.IDLE expected 1 but got 0
Harness: the test ran to completion.
